+++ src/fr/maxlego08/menu/api/utils/MetaUpdater.java
@@ -27,9 +27,9 @@ public interface MetaUpdater extends MessageSender {
     /**
      * Updates the display name of the {@link ItemMeta}.
      *
-     * @param itemMeta        The ItemMeta to be updated.
-     * @param text            The text to set as the display name.
-     * @param offlinePlayer   The player for whom MiniMessage will be applied.
+     * @param itemMeta      The ItemMeta to be updated.
+     * @param text          The text to set as the display name.
+     * @param offlinePlayer The player for whom MiniMessage will be applied.
      */
     void updateDisplayName(ItemMeta itemMeta, String text, OfflinePlayer offlinePlayer);
 
@@ -41,12 +41,13 @@ public interface MetaUpdater extends MessageSender {
      * @param player   The player for whom MiniMessage will be applied.
      */
     void updateLore(ItemMeta itemMeta, List<String> lore, Player player);
+
     /**
      * Updates the lore of the {@link ItemMeta}.
      *
-     * @param itemMeta        The ItemMeta to be updated.
-     * @param lore            The lore to set.
-     * @param offlinePlayer   The player for whom MiniMessage will be applied.
+     * @param itemMeta      The ItemMeta to be updated.
+     * @param lore          The lore to set.
+     * @param offlinePlayer The player for whom MiniMessage will be applied.
      */
     void updateLore(ItemMeta itemMeta, List<String> lore, OfflinePlayer offlinePlayer);
 
@@ -59,15 +60,24 @@ public interface MetaUpdater extends MessageSender {
      * @return Inventory with colored name
      */
     Inventory createInventory(String inventoryName, int size, InventoryHolder inventoryHolder);
+
     /**
      * Create an {@link Inventory}
      *
      * @param inventoryName   Inventory Name.
      * @param inventoryHolder Inventory Holder.
-     * @param inventoryType Inventory Type.
+     * @param inventoryType   Inventory Type.
      * @return Inventory with colored name
      */
     Inventory createInventory(String inventoryName, InventoryType inventoryType, InventoryHolder inventoryHolder);
 
+    /**
+     * Opens a book for the specified player.
+     *
+     * @param player The player who will view the book.
+     * @param title  The title of the book.
+     * @param author The author of the book.
+     * @param lines  The lines of the book.
+     */
     void openBook(Player player, String title, String author, List<String> lines);
 }