+++ changelog.md
@@ -40,10 +40,13 @@
 
 # Unreleased
 
+- Added inventory type [#97](https://github.com/Maxlego08/zMenu/pull/97)
+- Added placeholder priority [#96](https://github.com/Maxlego08/zMenu/pull/96)
 - Added the possibility to include a prefix for lists when using a pattern. You need to add text in front of your placeholder
 - Modified the ``money`` permissible to support more economies using [CurrenciesAPI](https://github.com/Traqueur-dev/CurrenciesAPI)
 - Modified the ``deposit`` and ``withdraw`` action to support more economies using [CurrenciesAPI](https://github.com/Traqueur-dev/CurrenciesAPI)
 - Added ``deny-message`` for commands. Allows to send a custom error message if the player does not have permission.
+- Fixed the max page placeholder that says 0, it will now say 1 by default
 
 # 1.0.3.6